Xresources, xinitrc, NERDTreeBookmarks
Xresources now creates an aesthetically pleasing configuration
for gruvbox with many modifications. Still a lot to be done but now
it has tabs, good looking font that renders at a reasonable size,
highlights that are easy to read etc.
Xinitrc explicitly sources .Xresources and .Xdefaults. I have no idea
in what situation startx doesn't already do that but figured I'd state
it.
NERDTreeBookmarks is a file I somehow manage to lose track of way too
often so it's going to be nice to have it in a repository.