I'm gonna be honest almost nothing works
Nvim highlights characters really  weird to the point of illegibility
The xinitrc and Xresource files need to be heavily modified so that
urxvt can work correctly

Also the languageclient plugin for nvim is acting erractically.
Genuinely unsure whats going on at any level.

i3's actually been okay but it's new enough that much more configuration
needs to happen to the status bar and it needs to accept native DM
panels because I still don't have a clear idea of what's on my
computer....

And oh boy isn't that just the beginning.