Working on snippet functionality.
The snippets plugin complains as soon as I enter Insert mode;
however, it still wrote my header.

It's either going to come to me squashing all warnings, deleting
all snippets out of the repo and hand piecing them back in to
ensure the syntax is correct, and dropping it entirely.

It keeps running into problems with parsing the headers, the snippets
themselves, and seemingly everything about these files.

Also tmux looks odd. But I feel good because nvim is currently a full
IDE with an autocompleter, a server that connects to asynchronous
linters with spell checking and seemless integration into the terminal
and file browsers. Also it starts in 400ms.