Modified almost everything in IPy dir. Matplotlib
Added the test file for IPy startup.
Added a new :mod:`01_rehashx.py` to prepend everything else in startup
Solves a huge problem with ipython aliasing. Should be in good condition
to use a system shell!

Added the git attributes file.
Deleted the TODO file.

Modified ~1000 lines in dircolors.
Bashrc now works with ag and fzf in a smoother manner.